Refactor generation to clean up code
