+++ docs/index.html
@@ -43,17 +43,6 @@
       .type_definition:nth-child(even) {
         border-color: #c4a776;
       }
-      .script_definition {
-        background-color: #00000012;
-        border-left-width: 4px;
-        border-left-style: solid;
-        border-color: #32abb877;
-        padding: 30px;
-        margin: 5px -30px;
-      }
-      .script_definition:nth-child(even) {
-        border-color: #d4fbff77;
-      }
     </style>
   </head>
   <body>
